fix(auth): extract isolated plugin-auth helper to pass max-lines
