+++ src/agents/embedded-agent-runner/run/auth-controller.test.ts
@@ -2,8 +2,7 @@
 import fs from "node:fs/promises";
 import os from "node:os";
 import path from "node:path";
-import type { Model } from "openclaw/plugin-sdk/llm";
-import { beforeEach, describe, expect, it, vi, type Mock } from "vitest";
+import { beforeEach, describe, expect, it, vi } from "vitest";
 import { createDeferred } from "../../../../test/helpers/promise.js";
 import { isSecretValueRegisteredForRedaction } from "../../../logging/secret-redaction-registry.js";
 import { SecretSurfaceUnavailableError } from "../../../secrets/runtime-degraded-state.js";
@@ -17,7 +16,6 @@ import { OAuthRefreshFailureError } from "../../auth-profiles/oauth-refresh-fail
 import { resolveAuthProfileOrder } from "../../auth-profiles/order.js";
 import { ensureAuthProfileStore, saveAuthProfileStore } from "../../auth-profiles/store.js";
 import { FailoverError } from "../../failover-error.js";
-import type { RuntimeAuthState } from "./helpers.js";
 
 const mocks = vi.hoisted(() => ({
   prepareProviderRuntimeAuth: vi.fn(),
@@ -45,133 +43,20 @@ vi.mock("../../model-auth.js", async () => {
 });
 
 import {
+  createMutableAuthControllerHarness,
+  createMutableEmbeddedRunAuthController,
+  createTestModel,
+  getRuntimeAuthSnapshot,
+} from "./auth-controller-harness.js";
+import {
   createEmbeddedRunAuthController,
   resolveEmbeddedAuthCooldownProbePolicy,
 } from "./auth-controller.js";
-
-function createTestModel(): Model {
-  return {
-    id: "test-model",
-    name: "test-model",
-    provider: "custom-openai",
-    api: "openai-responses",
-    baseUrl: "https://old.example.com/v1",
-    headers: {
-      Authorization: "Bearer stale-token",
-    },
-    reasoning: false,
-    input: ["text"],
-    cost: { input: 0, output: 0, cacheRead: 0, cacheWrite: 0 },
-    contextWindow: 8_000,
-    maxTokens: 4_000,
-  } as Model;
-}
-
-function getRuntimeAuthSnapshot(
-  state: RuntimeAuthState | null,
-): Pick<RuntimeAuthState, "profileId" | "refreshInFlight"> | null {
-  return state ? { profileId: state.profileId, refreshInFlight: state.refreshInFlight } : null;
-}
-
-type MutableAuthControllerHarness = {
-  runtimeModel: Model;
-  effectiveModel: Model;
-  apiKeyInfo: unknown;
-  lastProfileId?: string;
-  runtimeAuthState: RuntimeAuthState | null;
-  profileIndex: number;
-};
-
-type RuntimeApiKeySetter = Mock<(provider: string, apiKey: string) => void>;
 
 function expectProtectedRuntimeValue(value: string | undefined, plaintext: string): void {
   expect(value).not.toBe(plaintext);
   expect(looksLikeSecretSentinel(value ?? "")).toBe(true);
   expect(resolveSecretSentinel(value ?? "")).toBe(plaintext);
-}
-
-function createMutableAuthControllerHarness(): MutableAuthControllerHarness {
-  // Mutable harness mirrors the runner fields the auth controller updates
-  // through injected getters/setters.
-  return {
-    runtimeModel: createTestModel(),
-    effectiveModel: createTestModel(),
-    apiKeyInfo: null,
-    lastProfileId: undefined,
-    runtimeAuthState: null,
-    profileIndex: 0,
-  };
-}
-
-function createMutableEmbeddedRunAuthController(params: {
-  harness: MutableAuthControllerHarness;
-  setRuntimeApiKey: RuntimeApiKeySetter;
-  profileCandidates?: string[];
-  authStore?: AuthProfileStore;
-  fallbackConfigured?: boolean;
-  lockedProfileId?: string;
-  allowTransientCooldownProbe?: boolean;
-  warn?: (message: string) => void;
-  agentDir?: string;
-  prepareModelForAuthProfile?: Parameters<
-    typeof createEmbeddedRunAuthController
-  >[0]["prepareModelForAuthProfile"];
-}) {
-  return createEmbeddedRunAuthController({
-    config: undefined,
-    agentDir: params.agentDir ?? "/tmp/agent",
-    workspaceDir: "/tmp/workspace",
-    authStore:
-      params.authStore ??
-      ({
-        version: 1,
-        profiles: {},
-      } as AuthProfileStore),
-    authStorage: { setRuntimeApiKey: params.setRuntimeApiKey },
-    profileCandidates: params.profileCandidates ?? ["default"],
-    lockedProfileId: params.lockedProfileId,
-    initialThinkLevel: "medium",
-    attemptedThinking: new Set(),
-    fallbackConfigured: params.fallbackConfigured ?? false,
-    allowTransientCooldownProbe: params.allowTransientCooldownProbe ?? false,
-    getProvider: () => "custom-openai",
-    getModelId: () => "test-model",
-    getRuntimeModel: () => params.harness.runtimeModel,
-    setRuntimeModel: (next) => {
-      params.harness.runtimeModel = next;
-    },
-    getEffectiveModel: () => params.harness.effectiveModel,
-    setEffectiveModel: (next) => {
-      params.harness.effectiveModel = next;
-    },
-    getApiKeyInfo: () => params.harness.apiKeyInfo as never,
-    setApiKeyInfo: (next) => {
-      params.harness.apiKeyInfo = next;
-    },
-    getLastProfileId: () => params.harness.lastProfileId,
-    setLastProfileId: (next) => {
-      params.harness.lastProfileId = next;
-    },
-    getRuntimeAuthState: () => params.harness.runtimeAuthState as never,
-    setRuntimeAuthState: (next) => {
-      params.harness.runtimeAuthState = next;
-    },
-    getRuntimeAuthRefreshCancelled: () => false,
-    setRuntimeAuthRefreshCancelled: () => undefined,
-    getProfileIndex: () => params.harness.profileIndex,
-    setProfileIndex: (next) => {
-      params.harness.profileIndex = next;
-    },
-    ...(params.prepareModelForAuthProfile
-      ? { prepareModelForAuthProfile: params.prepareModelForAuthProfile }
-      : {}),
-    setThinkLevel: () => undefined,
-    log: {
-      debug: () => undefined,
-      info: () => undefined,
-      warn: params.warn ?? (() => undefined),
-    },
-  });
 }
 
 describe("createEmbeddedRunAuthController", () => {
