+++ src/agents/embedded-agent-runner/run/auth-controller.ts
@@ -24,11 +24,9 @@ import { FailoverError, resolveFailoverStatus } from "../../failover-error.js";
 import { shouldUseTransientCooldownProbeSlot } from "../../failover-policy.js";
 import { renderAuthProfileFailoverCopy } from "../../failover/user-copy.js";
 import {
-  createRuntimeProviderAuthLookup,
   getApiKeyForModelCore,
   MissingProviderAuthError,
   type ResolvedProviderAuth,
-  type RuntimeProviderAuthLookup,
 } from "../../model-auth.js";
 import { buildProviderAuthRecoveryHint } from "../../provider-auth-recovery-hint.js";
 import { providerModelRouteAcceptsAuthMode } from "../../provider-model-route-auth.js";
@@ -49,6 +47,7 @@ import {
   RUNTIME_AUTH_REFRESH_RETRY_MS,
   type RuntimeAuthState,
 } from "./helpers.js";
+import { createIsolatedPluginAuthResolver } from "./isolated-plugin-auth.js";
 import type { RunEmbeddedAgentParams } from "./params.js";
 
 type ApiKeyInfo = ResolvedProviderAuth;
@@ -507,25 +506,16 @@ export function createEmbeddedRunAuthController(params: {
     throw new Error(message);
   };
 
-  let runtimeAuthLookup: RuntimeProviderAuthLookup | undefined;
-  const resolveRuntimeAuthLookup = () =>
-    (runtimeAuthLookup ??= createRuntimeProviderAuthLookup({
-      cfg: params.config,
-      workspaceDir: params.workspaceDir,
-    }));
+  const isolatedPluginAuth = createIsolatedPluginAuthResolver({
+    cfg: params.config,
+    workspaceDir: params.workspaceDir,
+  });
 
   const resolveApiKeyForCandidate = async (
     candidate?: string,
     model = params.getRuntimeModel(),
     allowAuthProfileFallback?: boolean,
   ) => {
-    // Provider plugin synthetic-auth (e.g. GCP-ADC) is a provider-owned hook,
-    // not stored-profile discovery. An isolated direct attempt (prepared route
-    // disabled profile fallback) keeps it reachable on its own flag, with
-    // eligibility scoped by the prepared runtimeLookup. The ordinary profile
-    // path intentionally passes neither so its pre-existing unrestricted
-    // behavior is unchanged.
-    const isolatedDirectAttempt = allowAuthProfileFallback === false;
     return getApiKeyForModelCore({
       model,
       cfg: params.config,
@@ -535,9 +525,7 @@ export function createEmbeddedRunAuthController(params: {
       workspaceDir: params.workspaceDir,
       lockedProfile: candidate != null && candidate === params.lockedProfileId,
       allowAuthProfileFallback,
-      ...(isolatedDirectAttempt
-        ? { allowPluginSyntheticAuth: true, runtimeLookup: resolveRuntimeAuthLookup() }
-        : {}),
+      ...isolatedPluginAuth(allowAuthProfileFallback),
       secretSentinels: true,
     });
   };
